Add configuration filter to narrow deployment
Adds a -df -deployFilterfile configuration that is a file path to load
for telling the MavenRepositoryDeployer which GAVs to deploy (i.e.
filter in). This file contains a groupId:artifactId:version:ext
(patterns supported) per line, and is loaded into memory as a Set<Gav>.
It implements the Object.equals method on Gav and matcher support, so
that we can compare GAVs in order to know what to deploy. Only artifact
references that exist in the file will be deployed. If the filter file
is not specified, or doesnt exist, we do not filter at all. The
matching logic is made to be multi-threaded, in case the filter file is
large.

References: https://github.com/simpligility/maven-repository-tools/issues/81

Signed-off-by: Samuel Dacanay <sam.dacanay@chainguard.dev>